[SimpleRSS] bugfix & improvements
- bugfix: no more crashes when setting was set to 'ticker'
- bugfix: troubles when exiting plugin (the last screen before calling the plugin [=Pluginbrowser] was blank when exiting)
- improvement: all Python-files were integrated in the 'plugin.py'
- '**twisted.web.client.Agent.request(b'GET', uri)**' was replaced by '**twisted.internet.reactor.callInThread**' and '**requests.get(uri)**'
- translations modified accordingly
- Metrix-Skins have been overworked
